Created LibraryCard component to createa a unique state variable for each card. Refactored Library.js to render new component inside of map. Moved string related variables to StrManipulation.js.
