+++ src/components/library/Library.js
@@ -1,9 +1,9 @@
 import React, { useState , useEffect, useContext } from "react";
-import { Grid, Typography, Paper, IconButton, Button, Card, CardActionArea, CardActions, CardContent, CardMedia, Divider, List, ListItem, ListItemText, ListItemIcon, Collapse } from "@material-ui/core";
-import { Favorite, Search, Shuffle, LocalLibrary, Subject, Label, ExpandLess, ExpandMore} from "@material-ui/icons";
+import { Grid, Typography, Paper } from "@material-ui/core";
+import { Favorite, Search, Shuffle, LocalLibrary } from "@material-ui/icons";
 import { makeStyles } from "@material-ui/core/styles";
 import { LibraryContext } from "./LibraryProvider";
-import clsx from "clsx";
+import { LibraryCard } from "./LibraryCard";
 
 // Declare variable to import material-ui components and specify local theme overrides 
 const useStyles = makeStyles((theme) => ({
@@ -19,91 +19,6 @@ const useStyles = makeStyles((theme) => ({
         flexWrap: 'wrap',
         justifyContent: 'center',
     },
-    cardStyle: {
-        // Primary theme for rendered game cards
-        width: theme.spacing(40),
-        height: 600,
-        margin: theme.spacing(2),
-        backgroundColor: theme.palette.primary.light,
-        color: '#f5f5f5',
-        transition: "0.3s",
-        boxShadow: "0 8px 40px -12px rgba(0,0,0,0.3)",
-        "&:hover": {
-            boxShadow: "0 16px 70px -12.125px rgba(0,0,0,0.3)"
-        },
-    },
-    cardHeaderSpan: {
-        // Set default height for card header
-        marginTop: theme.spacing(1),
-        height: theme.spacing(10),
-        textAlign: 'center',
-    },
-    cardHeader: {
-        fontWeight: "bold",
-    },
-
-    cardMedia: {
-        // Default styles for card images 
-        paddingTop: "15.5%",
-        height: 160,
-        width: '100%',
-    },
-    cardButtonContainer: {
-        // Container to anchor card buttons to bottom
-        display: 'grid',
-        height: theme.spacing(12),
-        alignContent: 'end',
-        alignItems: 'end',
-    },
-    cardDivider: {
-        margin: `${theme.spacing(3)}px 0`
-    },
-    margin: {
-        margin: theme.spacing(2),
-    },
-    genreContainer: {
-        display: 'flex',
-        justifyContent: 'center',
-        width: '100%',
-    },
-    genreTag: {
-        display: 'flex',
-        flexWrap: 'wrap',
-        justifyContent: 'start',
-        alignItems: 'center',
-        minHeight: theme.spacing(5),
-        minWidth: '40%',
-        maxWidth: '100%',
-        border: `.1rem solid #f5f5f5`,
-        backgroundColor: theme.palette.secondary.light,
-        borderRadius: '.5rem',
-        fontWeight: 300,
-        marginBottom: theme.spacing(1),
-    },
-    uncategorizedTag: {
-        display: 'flex',
-        flexWrap: 'wrap',
-        justifyContent: 'start',
-        alignItems: 'center',
-        minHeight: theme.spacing(5),
-        width: theme.spacing(20),
-        border: `.1rem solid #f5f5f5`,
-        backgroundColor: theme.palette.secondary.light,
-        borderRadius: '.5rem',
-        fontWeight: 300,
-        marginBottom: theme.spacing(1),
-    },
-    tagText: {
-        marginLeft: theme.spacing(1),
-        marginRight: theme.spacing(2),
-        fontSize: 15,
-        textAlign: 'left',
-    },
-    tagIcon: {
-        marginLeft: theme.spacing(2),
-        fontSize: 20,
-        textAlign: 'left',
-    },
     paper: {
         // Rectanglar blue-gray div
         width: theme.spacing(100),
@@ -112,9 +27,6 @@ const useStyles = makeStyles((theme) => ({
         margin: theme.spacing(3),
         backgroundColor: theme.palette.secondary.main,
         color: '#f5f5f5',
-    },
-    nested: {
-        paddingLeft: theme.spacing(4),
     },
     iconText: {
         // Alternate text alignment
@@ -205,64 +117,12 @@ export const LibraryList = () => {
     return (
         <>
         <div className={classes.cardContainer}>
-
-        {allGames.map(game => {
-                // const cardTitle = game.title.replace(/ *\([^)]*\) */g, '').replace('Internet Arcade:', '')
-                const cardTitle = (title, length) => {
-                    let editedTitle = title.replace(/ *\([^)]*\) */g, '').replace('Internet Arcade:', '')
-                    return truncate(editedTitle, length)
-                };
-                return <>
-                <Card className={classes.cardStyle}>
-                    <CardMedia
-                    className={classes.cardMedia}
-                    image={`https://archive.org/services/img/${game.identifier}`}
-                    title={cardTitle(game.title, 44)}
-                    />
-                    <CardContent>
-                        <div className={classes.cardHeaderSpan}>
-                            <Typography className={classes.cardHeader}  variant="h5" component="h2" gutterBottom>
-                                {cardTitle(game.title, 41)}
-                            </Typography>
-                        </div>
-                        <div className={classes.genreContainer}>
-                            {!game.genre ? <div className={classes.uncategorizedTag}>
-                                <Label className={classes.tagIcon} /><Typography className={classes.tagText} variant="body1" component="p">Uncategorized</Typography>
-                            </div>:<div className={classes.genreTag}>
-                                <Label className={classes.tagIcon} /><Typography className={classes.tagText} variant="body1" component="p">{truncate(game.genre, 22)}</Typography>
-                            </div>}
-                        </div>
-                        <Divider className={classes.cardDivider} />
-                        <List component="nav" disablePadding>
-                            <ListItem button onClick={handleCardDetails}>
-                                <ListItemIcon>
-                                    <Subject />
-                                </ListItemIcon>
-                                <ListItemText primary="Details" />
-                                {open ? <ExpandLess /> : <ExpandMore />}
-                            </ListItem>
-                            <Collapse in={open} timeout="auto" unmountOnExit>
-                                <List component="div" disablePadding>
-                                    <ListItem className={classes.nested}>
-                                        <ListItemText primary="Release Year" secondary={game.date} />
-                                    </ListItem>
-                                    <ListItem className={classes.nested}>
-                                        <ListItemText primary="Publisher" secondary={game.creator} />
-                                    </ListItem>
-                                </List>
-                            </Collapse>
-                        </List>
-                    </CardContent>
-                    <CardActions>
-                        <div className={classes.cardButtonContainer}>
-                            <IconButton>
-                                <Favorite className={classes.redIcon}/>
-                            </IconButton>
-                        </div>
-                    </CardActions>
-                </Card>
-            </>         
-})}
+            {allGames.map(game => {
+                
+                return  <>
+                            <LibraryCard key={`${game.indentifier}--card`} game={game}/>
+                        </>         
+            })}
         </div>
         </>
