+++ src/components/library/LibraryProvider.js
@@ -6,9 +6,10 @@ export const LibraryProvider = (props) => {
     const [ allGames , setAllGames ] = useState([])
 
     const getArcadeTitles  = () => {
-        return fetch(`https://archive.org/services/search/v1/scrape?fields=title,genre,identifier,date,creator&q=collection%3Ainternetarcade`)
+        return fetch(`https://archive.org/services/search/v1/scrape?fields=title,genre,identifier,date,creator&q=collection%3Ainternetarcade&count=200`)
         .then(r => r.json())
         .then(searchResults => {
+            // console.log(searchResults);            
             setAllGames(searchResults.items)
         })
     };
